refactor(api): simplify directory structure of context providers
