+++ packages/api/src/config/context/ComponentRegistrar.ts
